+++ shelter/__init__.py
@@ -3,4 +3,4 @@ Simple Python's Tornado wrapper which provides helpers for creating
 a new project, writing management commands, service processes, ...
 """
 
-__version___ = '1.1.5'
+__version___ = '1.2.0.dev1'